Refactor components folders structure
