+++ woolly/src/components/sales/SaleCard.jsx
@@ -4,7 +4,7 @@ import { shorten } from '../../utils';
 
 import { withStyles } from '@material-ui/core/styles';
 import { Card, CardContent, CardActions } from '@material-ui/core';
-import { NavButton } from './Nav.jsx';
+import { NavButton } from '../common/Nav';
 
 
 class SaleCard extends React.Component {
