+++ woolly/src/pages/Sales.jsx
@@ -3,8 +3,8 @@ import { connect } from 'react-redux';
 import actions from '../redux/actions';
 import { withStyles } from '@material-ui/core/styles';
 
-import Loader from '../components/Loader';
-import SaleCard from '../components/common/SaleCard';
+import Loader from '../components/common/Loader';
+import SaleCard from '../components/sales/SaleCard';
 
 const connector = connect(store => ({
 	sales: store.getData('sales', []),
